ai_service/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
anthropic==0.25.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.0
